Remove irrelevant transcripts.py
Several assumptions have changed: there is no Dragon NaturallySpeaking,
the filenames on the recordings are expected to be useful rather than
generic, and the QFile app seems to preserve mtimes anyway.

As such, there are a few problems with the script: there is no need for
any requirements.txt (no RTF or audio manipulation in Python) and the
submodule pattern is too ambitious (due to Audacity doing most of the
work). Remaining file manipulation is more easily done with sh.